Extract permissions logic into its own module. Address css comments.
